Extract mix and merge into lib/utils.
