+++ lib/cluster/base.js
@@ -4,6 +4,7 @@ ComboHandler.Cluster.Base
 var EventEmitter = require('events').EventEmitter;
 
 var args = require('../args');
+var merge = require('../utils').merge;
 
 exports = module.exports = ComboBase;
 
@@ -73,25 +74,3 @@ ComboBase.prototype.listen = function (port) {
 
     return this.start(cb.bind(this));
 };
-
-// -- Private Methods ---------------------------------------------------------
-
-function mix(receiver, supplier) {
-    if (supplier) {
-        Object.keys(supplier).forEach(function (k) {
-            receiver[k] = supplier[k];
-        });
-    }
-
-    return receiver;
-}
-
-function merge() {
-    var result = {};
-
-    ([].slice.call(arguments)).forEach(function (obj) {
-        mix(result, obj);
-    });
-
-    return result;
-}
